+++ index.js
@@ -21,12 +21,14 @@ const rotasClientes = require("./routes/clientes");
 const rotasPets = require("./routes/pets");
 const rotasProdutos = require("./routes/produtos");
 const rotasServico = require("./routes/servicos");
+const rotasAgendamentos = require("./routes/agendamentos")
 
 // Juntar ao app as rotas dos arquivos
 app.use(rotasClientes); // Configurar o grupo de rotas no app
 app.use(rotasPets);
 app.use(rotasProdutos);
 app.use(rotasServico);
+app.use(rotasAgendamentos);
 
 // Escuta de eventos (listen)
 app.listen(3001, () => {
